Set check_mode=no for task ssh-keyscan
- name: "Scan and register SSH host keys based on hostname"
  command: "ssh-keyscan {{ item }}"
  register: key_scan_name
  changed_when: false
  loop: "{{ hostvars | flatten(levels=1) }}"

- name: "Append SSH host keys based on hostname to list"
  set_fact:
    ssh_host_keys: "{{ ssh_host_keys }} + [ '{{ item.1 }}' ]"
  loop: "{{ key_scan_name['results']|subelements('stdout_lines') }}"
  loop_control:
    label: "{{ item.1.split()[0] }} {{ item.1.split()[1] }}"

If Ansible runs in check mode, the first task is skipped while the
second one fails because the key_scan_name variable doesn’t have a
list of ssh keys to iterate over. The reason why this happens is because
shell and command modules don’t support check mode, so they are skipped.
To overcome this problem, we can just add check_mode: no to the shell
task. check_mode: no basically means never execute a task in check mode,
which translates to always run it regularly.

So it all looks like this:

- name: "Scan and register SSH host keys based on hostname"
  command: "ssh-keyscan {{ item }}"
  register: key_scan_name
  changed_when: false
  check_mode: no
  loop: "{{ hostvars | flatten(levels=1) }}"

- name: "Append SSH host keys based on hostname to list"
  set_fact:
    ssh_host_keys: "{{ ssh_host_keys }} + [ '{{ item.1 }}' ]"
  loop: "{{ key_scan_name['results']|subelements('stdout_lines') }}"
  loop_control:
    label: "{{ item.1.split()[0] }} {{ item.1.split()[1] }}"